Added a bunch of stuff, check description
Flags are able to be set and unset.
Game info is displayed at the bottom.
Autoclicking for blank tiles is faster, although still not optimal. For
optimal performance, I should put checked tiles into a separate group,
I'm just too lazy.
Winning or losing displays a message in the middle of the grid.
New game can be started after winning or losing.